+++ R/gcd.R
@@ -1,0 +1,4 @@
+gcd <- function(x,y) {
+	r <- x%%y;
+	return(ifelse(r, gcd(y, r), y))
+}